test: move tests into premium
Adding project approval rules and setting code_owner_approvals
 or allow_force_push require premium license
