+++ tests/acceptance/standard/test_branches.py
@@ -1,7 +1,4 @@
-import logging
-
-from gitlab import GitlabGetError, GitlabCreateError
-from gitlab.v4.objects import ProjectProtectedBranch
+from gitlab import GitlabGetError
 
 from gitlabform.gitlab import AccessLevel
 from tests.acceptance import get_only_branch_access_levels, run_gitlabform
@@ -51,65 +48,6 @@ class TestBranches:
         protected_branches = project.protectedbranches.list()
         for pb in protected_branches:
             assert pb.name != branch
-
-    def test__modify_force_push_and_code_owner_approvals(self, project, branch):
-        try:
-            protected_branch = project.protectedbranches.get(branch)
-            protected_branch.delete()
-        except GitlabGetError:
-            # Branch currently not protected so do nothing
-            logging.debug("Nothing to reset")
-
-        config_protect_branch = f"""
-        projects_and_groups:
-          {project.path_with_namespace}:
-            branches:
-              {branch}:
-                protected: true
-                allow_force_push: false
-                code_owner_approval_required: true
-        """
-
-        run_gitlabform(config_protect_branch, project.path_with_namespace)
-
-        protected_branch = project.protectedbranches.get(branch)
-        assert protected_branch.allow_force_push is False
-        assert protected_branch.code_owner_approval_required is True
-
-        config_modify_settings_for_branch = f"""
-        projects_and_groups:
-          {project.path_with_namespace}:
-            branches:
-              {branch}:
-                protected: true
-                allow_force_push: true
-                code_owner_approval_required: false
-        """
-
-        run_gitlabform(config_modify_settings_for_branch, project.path_with_namespace)
-
-        protected_branch = project.protectedbranches.get(branch)
-        assert protected_branch.allow_force_push is True
-        assert protected_branch.code_owner_approval_required is False
-
-        config_remove_settings_for_branch = f"""
-        projects_and_groups:
-          {project.path_with_namespace}:
-            branches:
-              {branch}:
-                protected: true
-        """
-
-        run_gitlabform(config_remove_settings_for_branch, project.path_with_namespace)
-
-        # If we were protecting a branch for the first time and not passing these values, Gitlab would default both
-        # to "False" as per: https://docs.gitlab.com/api/protected_branches/#protect-repository-branches
-        # We retain that decision-making rather than leaving the values at their previous state, otherwise
-        # if a User manually unprotected a branch, and restored protection by re-running Gitlabform, the flags would
-        # have silently changed
-        protected_branch = project.protectedbranches.get(branch)
-        assert protected_branch.allow_force_push is False
-        assert protected_branch.code_owner_approval_required is False
 
     def test__repeatedly_modify_protection_rules(self, project_for_function, branch_for_function):
         """
@@ -218,99 +156,6 @@ class TestBranches:
         assert push_access_user_ids == []
         assert merge_access_user_ids == []
         assert unprotect_access_level is AccessLevel.MAINTAINER.value
-
-    def test__if_protected_branch_config_does_not_change_then_branch_approval_rules_are_retained(
-        self, project_for_function, branch_for_function
-    ):
-        # Previously we have unprotected and re-protected branches in order to apply config changes
-        # and even if config did not change, branches_processor would thing it still needed a change.
-        # This resulted in loss of manual approval rules: https://github.com/gitlabform/gitlabform/issues/1061
-        # This test can check that branch protection is not modified when the config is unchanged between glf runs
-        # when run in debug mode with breakpoints from an IDE, and validates the branch approval rules being retained
-        config_protect_branch = f"""
-         projects_and_groups:
-           {project_for_function.path_with_namespace}:
-             branches:
-               {branch_for_function}:
-                 protected: true
-                 push_access_level: {AccessLevel.NO_ACCESS.value}
-                 merge_access_level: {AccessLevel.MAINTAINER.value}
-                 unprotect_access_level: {AccessLevel.MAINTAINER.value}
-         """
-
-        run_gitlabform(config_protect_branch, project_for_function.path_with_namespace)
-
-        (
-            push_access_levels,
-            merge_access_levels,
-            push_access_user_ids,
-            merge_access_user_ids,
-            _,
-            _,
-            unprotect_access_level,
-        ) = get_only_branch_access_levels(project_for_function, branch_for_function)
-        assert push_access_levels == [AccessLevel.NO_ACCESS.value]
-        assert merge_access_levels == [AccessLevel.MAINTAINER.value]
-        assert push_access_user_ids == []
-        assert merge_access_user_ids == []
-        assert unprotect_access_level is AccessLevel.MAINTAINER.value
-
-        # Add manual approval rule on the protected branch
-        protected_branch: ProjectProtectedBranch = project_for_function.protectedbranches.get(branch_for_function)
-        try:
-            project_for_function.approvalrules.create(
-                {
-                    "name": "Branch Protection validation",
-                    "approvals_required": 2,
-                    "rule_type": "regular",
-                    "protected_branch_ids": [
-                        protected_branch.id,
-                    ],
-                }
-            )
-        except GitlabCreateError as e:
-            logging.error(f"Error creating approval rules {e.error_message}")
-            raise e
-
-        approval_rules = project_for_function.approvalrules.list(get_all=True)
-        assert len(approval_rules) == 1
-        approval_rule = approval_rules[0]
-        assert approval_rule.name == "any"
-        assert approval_rule.approvals_required == 2
-        assert len(approval_rule.protected_branches) == 1
-        pb_ar = approval_rule.protected_branches[0]
-        assert pb_ar.get("id") == protected_branch.id
-
-        # re-run the exact same config
-        run_gitlabform(config_protect_branch, project_for_function.path_with_namespace)
-
-        (
-            push_access_levels,
-            merge_access_levels,
-            push_access_user_ids,
-            merge_access_user_ids,
-            _,
-            _,
-            unprotect_access_level,
-        ) = get_only_branch_access_levels(project_for_function, branch_for_function)
-
-        assert push_access_levels == [AccessLevel.NO_ACCESS.value]
-        assert merge_access_levels == [AccessLevel.MAINTAINER.value]
-        assert push_access_user_ids == []
-        assert merge_access_user_ids == []
-        assert unprotect_access_level is AccessLevel.MAINTAINER.value
-
-        # If the branch was unprotected and then re-protected the branch id of the protected branch would likely differ
-        # from that stored in the branch approval rule
-        protected_branch: ProjectProtectedBranch = project_for_function.protectedbranches.get(branch_for_function)
-
-        approval_rules = project_for_function.approvalrules.list(get_all=True)
-        assert len(approval_rules) == 1
-        approval_rule = approval_rules[0]
-        assert approval_rule.name == "any"
-        assert approval_rule.approvals_required == 2
-        pb_ar = approval_rule.protected_branches[0]
-        assert pb_ar.get("id") == protected_branch.id
 
     def test__protect_with_wildcard(self, project):
         branch_wildcard = "r-*"
